cmd: Move container image promotion commands into kpromo
