+++ cmd/kpromo/cmd/root.go
@@ -22,6 +22,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
+	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/cip"
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/manifest"
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/run"
 	"sigs.k8s.io/release-utils/log"
@@ -71,6 +72,7 @@ func init() {
 	)
 
 	rootCmd.AddCommand(run.RunCmd)
+	rootCmd.AddCommand(cip.CipCmd)
 	rootCmd.AddCommand(manifest.ManifestCmd)
 }
